Reduced mocking in the journey responder tests
Now testing the parser functionality directly, rather than by inspecting
the side effect function calls. This makes this feature much easier to
implement.
